+++ application/views/actor.php
@@ -1,7 +1,9 @@
 <div class="row">
     <div class="span4">
+            <label>{who} plays {role}</label>
             <a href='/welcome/show/{key}'>
-                    <img src="/pix/{who}.jpg" title="{who} plays {role}"/>
-            </a>
+                    <img src="/pix/{who}.jpg"/>
+            </a>            
+            <label>{what}</label>
     </div>
 </div>